+++ src/main/java/vs/representative/features/corrupted/votes/CorruptedVotesController.java
@@ -2,15 +2,12 @@ package vs.representative.features.corrupted.votes;
 
 import java.util.List;
 
-import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -50,7 +47,7 @@ public class CorruptedVotesController {
 	@ResponseStatus(org.springframework.http.HttpStatus.OK)
 	@ApiOperation(value = "Get invalid votes by id")
 	public Long getCorruptVotesById(@PathVariable("id") Integer id) {
-		return corruptedVotesRepository.getCorruptedVotesInConstituency(id);
+		return corruptedVotesRepository.getCorruptedVotesInConstituency(id).longValue();
 	}
